Initial commit
This is a skeleton for an XMLTV grabber that works on `zap2it.com`.

Currently, option parsing for the requisite XMLTV parameters is
implemented, as well as the absolute base requirements for an XMLTV
grabber to identify itself and its capabilities.

Additionally, a switch to print lineup info has been implemented using a
`fetch` framework for making future cache-aware API requests to the
appropriate host.