Unify caching
We now create a persistent cache object at launch and use it throughout
the program instead of opening a file on every fetch.

This way, we serialize and de-serialize the cache a once at load then
only once again at program exit. Additionally, if no caching is
specified, the program will at least use a shared in-memory cache, even
if this cache isn't persisted to disk.