Stop overuse of `OpenStruct` to speed things up
Using plain hashes (with symbols though, we're not animals!) makes the
whole program run a _lot_ faster, especially loading and storing the
cache.

As an additional benefit, we now use `Hash#fetch` pretty extensively,
which explodes if it can't find the expected key. This way, we'll fail
before passing a `nil` somewhere down the line!