Use UTC, fix date ranging, add rate limiting
We now use UTC everywhere. Ideally, we'd convert on input and output,
but that can come later.

This also fixes date range calculating when downloading the lineup,
ensuring that there aren't off-by-one errors and the like.

Additionally, we add a per-domain rate limit of two requests per second
to ensure that we're not overwhelming the hosts from which we're
downloading data.