Fix channel name output, add another name
The "fancy name" for a channel might not exists, so we only add it when
it does.

We also add another "more canonical" name for the channel, one that
contains both its number and its call sign together. This will hopefully
help Tvheadend better map existing channels to those from the XMLTV
source.