Change a display name for better tvheadend compat
Putting a number first causes tvheadend to attempt to use that
`display-name` tag as the channel number, but since the value in the
name format "{number} {call-sign}" we were using _isn't_ a real channel
number, the heuristic would fail and we wouldn't get a mapped channel
number at all.

This changes the display name to put the number _last_ instead of first
so that it hopefully won't trigger that heuristic anymore, and the
number-only `display-name` will get picked up as the channel number
instead.